attempt to get the docker image of resolve running
