+++ app/resolve/Dockerfile.cli
@@ -6,4 +6,4 @@ RUN pip install poetry
 
 RUN poetry install
 
-CMD [ "resolve-cli" ]
+CMD ["poetry", "run", "resolve-cli"]
