+++ app/resolve/core/constants.py
@@ -1,7 +1,7 @@
-from rquest_omop_worker.enums import DistributionQueryType
+from core.enums import DistributionQueryType
 
 
 DISTRIBUTION_TYPE_FILE_NAMES_MAP = {
     DistributionQueryType.DEMOGRAPHICS: "demographics.distribution",
-    DistributionQueryType.GENERIC: "code.distribution"
+    DistributionQueryType.GENERIC: "code.distribution",
 }
