+++ app/resolve/core/rquest_dto/cohort.py
@@ -1,13 +1,13 @@
 from typing import List
-from rquest_omop_worker.rquest_dto.base_dto import BaseDto
-from rquest_omop_worker.rquest_dto.group import Group
+from core.rquest_dto.base_dto import BaseDto
+from core.rquest_dto.group import Group
 
 
 class Cohort(BaseDto):
     def __init__(self, groups: List[Group], groups_operator: str) -> None:
         self.groups = groups
         self.groups_operator = groups_operator
-    
+
     def to_dict(self) -> dict:
         """Convert `Cohort` to `dict`
