+++ app/resolve/core/rquest_dto/query.py
@@ -1,6 +1,6 @@
-from rquest_omop_worker.enums import DistributionQueryType
-from rquest_omop_worker.rquest_dto.base_dto import BaseDto
-from rquest_omop_worker.rquest_dto.cohort import Cohort
+from core.enums import DistributionQueryType
+from core.rquest_dto.base_dto import BaseDto
+from core.rquest_dto.cohort import Cohort
 
 
 class AvailabilityQuery(BaseDto):
